+++ lib/src/repo/remote_config/remote_config_repository.dart
@@ -6,7 +6,8 @@ import 'package:impaktfull_architecture/src/util/connectivity/connectivity.dart'
 import 'package:impaktfull_architecture/src/util/logger/logger.dart';
 import 'package:impaktfull_architecture/src/util/storage/shared_prefs/shared_prefs_store.dart';
 
-abstract class ImpaktfullRemoteConfigRepository<T extends ImpaktfullRemoteConfigData> {
+abstract class ImpaktfullRemoteConfigRepository<
+    T extends ImpaktfullRemoteConfigData> {
   final SharedPrefsStore _sharedPrefsStore;
   final ConnectivityController _connectivityController;
   final String storageKey;